requirements.txt
python-telegram-bot==21.9
Telethon==1.38.1